Make 64-bit VS compile clean. Mostly size_t vs. int tweaks.
git-svn-id: https://cvs.khronos.org/svn/repos/ogl/trunk/ecosystem/public/sdk/tools/glslang@25411 e7fa87d3-cd2b-0410-9028-fcbf551c1848

+++ StandAlone/StandAlone.cpp
@@ -760,7 +760,7 @@ void CompileFile(const char* fileName, ShHandle compiler)
 
     // move to length-based strings, rather than null-terminated strings
     for (int s = 0; s < NumShaderStrings; ++s)
-        lengths[s] = strlen(shaderStrings[s]);
+        lengths[s] = (int)strlen(shaderStrings[s]);
 
     if (! shaderStrings) {
         CompileFailed = true;
